refactor: decompose Calendar_Query into focused modules (phase 1)
Extract 7 focused modules from the 1,145-line Calendar_Query god class:

- Cache/CalendarCache — cache key generation, transient get/set, TTLs
- Data/EventHydrator — block parsing, meta/taxonomy hydration
- Display/DisplayVars — time formatting, unicode decode, display vars
- Display/EventRenderer — HTML rendering, templates, lazy placeholders
- Grouping/DateGrouper — date grouping, multi-day expansion, gap detection
- Grouping/MultiDayResolver — multi-day detection, date range generation
- Pagination/PageBoundary — unique dates, page boundaries, event counting
- Query/EventQueryBuilder — WP_Query args, meta/tax queries, geo filter

Calendar_Query becomes a thin facade delegating to modules.
Zero external API changes — CalendarAbilities continues to work
unchanged. Phase 2 will wire CalendarAbilities directly to modules.

Cache_Invalidator updated to use CalendarCache::PREFIX.

Closes #49
