+++ inc/Blocks/Calendar/Cache_Invalidator.php
@@ -12,6 +12,7 @@
 namespace DataMachineEvents\Blocks\Calendar;
 
 use DataMachineEvents\Core\Event_Post_Type;
+use DataMachineEvents\Blocks\Calendar\Cache\CalendarCache;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -71,8 +72,8 @@ class Cache_Invalidator {
 		$wpdb->query(
 			$wpdb->prepare(
 				"DELETE FROM {$wpdb->options} WHERE option_name LIKE %s OR option_name LIKE %s",
-				'_transient_' . Calendar_Query::CACHE_PREFIX . '%',
-				'_transient_timeout_' . Calendar_Query::CACHE_PREFIX . '%'
+				'_transient_' . CalendarCache::PREFIX . '%',
+				'_transient_timeout_' . CalendarCache::PREFIX . '%'
 			)
 		);
